+++ src/components/LMarkerClusterGroup.vue
@@ -1,19 +1,25 @@
 <script lang="ts">
-import type { MarkerClusterGroup } from 'leaflet'
+import type { MarkerClusterGroup, LeafletEventHandlerFnMap } from 'leaflet'
 import { ref, markRaw, onMounted, nextTick } from 'vue'
 import 'leaflet.markercluster'
 import 'leaflet.markercluster/dist/MarkerCluster.css'
 import 'leaflet.markercluster/dist/MarkerCluster.Default.css'
 import { InjectionKeys, Functions, Utilities } from '@vue-leaflet/vue-leaflet'
 
-import { markerClusterGroupProps, setupMarkerClusterGroup } from '@/functions/markerClusterGroup'
+import type { LeafletEventKeys, EventHandlerFn, EmitterEvents } from '@/types/markercluster'
+import {
+  markerClusterGroupProps,
+  markerClusterGroupEvents,
+  setupMarkerClusterGroup
+} from '@/functions/markerClusterGroup'
 
-const { remapEvents, propsBinder, assertInject, WINDOW_OR_GLOBAL } = Utilities
+const { propsBinder, assertInject, WINDOW_OR_GLOBAL } = Utilities
 const { AddLayerInjection } = InjectionKeys
 const { render } = Functions.Layer
 
 export default {
   props: markerClusterGroupProps,
+  emits: ['ready', ...markerClusterGroupEvents] as string[],
   setup(props, context) {
     const leafletObject = ref<MarkerClusterGroup>()
     const ready = ref(false)
@@ -25,9 +31,20 @@ export default {
     onMounted(async () => {
       const { markerClusterGroup } = WINDOW_OR_GLOBAL.L
       leafletObject.value = markRaw(markerClusterGroup(options))
+      const emitter =
+        (key: LeafletEventKeys[number]): EventHandlerFn =>
+        (e: EmitterEvents) =>
+          context.emit(key, e)
 
-      const { listeners } = remapEvents(context.attrs)
-      leafletObject.value.on(listeners)
+      const remapEvents = (): LeafletEventHandlerFnMap => {
+        const listeners: LeafletEventHandlerFnMap = {}
+        for (const event of markerClusterGroupEvents) {
+          listeners[event] = emitter(event)
+        }
+
+        return listeners
+      }
+      leafletObject.value?.on(remapEvents())
 
       propsBinder(methods, leafletObject.value, props)
       addLayer({
@@ -46,3 +63,4 @@ export default {
   }
 }
 </script>
+@/types/markercluster