+++ src/playground/App.vue
@@ -1,8 +1,12 @@
 <template>
   <div class="main-wrapper">
     <div class="menu">
-      <router-link to="/">Home</router-link>
-      <router-link to="/10000-marker">10.000 Marker</router-link>
+      <v-list density="compact" nav>
+        <v-list-item title="Home" to="/" />
+        <v-list-item title="Marker props" to="/props-marker" />
+        <v-list-item title="Marker events" to="/event-marker" />
+        <v-list-item title="10.000 Marker" to="/10000-marker" />
+      </v-list>
     </div>
     <div class="map-wrapper">
       <router-view />
@@ -19,10 +23,13 @@ import 'leaflet/dist/leaflet.css'
 <style>
 .main-wrapper {
   display: flex;
+  max-width: 1000px;
+  margin: 0 auto;
 }
 .menu {
   display: flex;
   flex-direction: column;
+  width: 150px;
 }
 
 .menu a {
@@ -35,7 +42,11 @@ import 'leaflet/dist/leaflet.css'
   background-color: lightgray;
 }
 .map-wrapper {
-  height: 75vh;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  padding: 8px;
+  gap: 8px;
   width: 100%;
 }
 </style>