+++ src/playground/views/MarkerDemo.10000.vue
@@ -1,5 +1,5 @@
 <template>
-  <div style="height: 600px; width: 800px">
+  <div style="height: 50vh; width: 100%">
     <l-map :useGlobalLeaflet="true" :zoom="13" :center="[-37.9, 175.46]">
       <l-tile-layer
         url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"